# docs/openapi/v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Eldercare Edge Enrollment and Topology API",
    "version": "1.0.0",
    "description": "Canonical frozen v1 contract for edge credential administration, installation enrollment, edge-authored topology reconciliation, machine ingest/config access, and audited clip download. Server-derived facility, installation, and enrollment-generation identity is authoritative on every edge-bearer request."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin API server"
    }
  ],
...
  "paths": {
...
    "/api/v1/alerts": {
      "get": {
        "description": "Returns facility-scoped alerts with optional status, sequence, and limit filters for dashboard history and reconciliation.",
        "operationId": "listAlerts",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "NEW",
                "ACKED",
                "RESOLVED"
              ],
              "type": "string"
            }
          },
          {
            "name": "afterSeq",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            }
          },
          {
            "name": "beforeSeq",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertList"
                }
              }
            }
          }
        },
        "security": [
          {
            "app_session": []
          }
        ],
        "summary": "List alerts",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/api/v1/alerts/{id}": {
      "get": {
        "description": "Returns a single facility-scoped alert by id or 404 when it is outside the caller's facility.",
        "operationId": "getAlert",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "app_session": []
          }
        ],
        "summary": "Get one alert",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/api/v1/alerts/{id}/resolve": {
      "patch": {
        "description": "Marks an alert resolved by the current user and emits the live alert-updated dashboard frame.",
        "operationId": "resolveAlert",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alert"
                }
              }
            }
          }
        },
        "security": [
          {
            "app_session": []
          }
        ],
        "summary": "Resolve an alert",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/api/v1/dashboard/stream": {
      "get": {
        "description": "Opens the facility-scoped SSE stream that emits only room-centric alert and alert-updated frames plus session-invalid control frames.",
        "operationId": "streamDashboardAlerts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "app_session": []
          }
        ],
        "summary": "Stream live dashboard alerts",
        "tags": [
          "DashboardStream"
        ],
        "x-sse-event-schemas": {
          "alert": "#/components/schemas/AlertSseData",
          "alert-updated": "#/components/schemas/AlertUpdatedSseData"
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "app_session",
        "description": "HttpOnly product session. Operation metadata further restricts the required role or capability."
      },
      "edgeBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "eft_v1.{tokenId}.{one-time-value}",
        "description": "Installation- and enrollment-generation-bound v1 edge principal. A legacy shared token is not authorized by this security scheme."
      }
    },
...
    "schemas": {
...
      "AlertActor": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string"
          }
        },
        "required": [
          "nickname"
        ]
      },
      "AlertSpace": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "Alert": {
        "type": "object",
        "properties": {
          "alertSeq": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "id": {
            "type": "string"
          },
          "backendEventId": {
            "type": "string"
          },
          "facilityId": {
            "type": "string"
          },
          "cameraId": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "spaceId": {
            "type": "string"
          },
          "room": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "probability": {
            "type": "number"
          },
          "snapshotKey": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "detectedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "ACKED",
              "RESOLVED"
            ]
          },
          "ackedById": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "ackedAt": {
            "oneOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "ackedBy": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AlertActor"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolvedById": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolvedAt": {
            "oneOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolvedBy": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AlertActor"
              },
              {
                "type": "null"
              }
            ]
          },
          "space": {
            "$ref": "#/components/schemas/AlertSpace"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "alertSeq",
          "id",
          "backendEventId",
          "facilityId",
          "cameraId",
          "spaceId",
          "room",
          "type",
          "probability",
          "snapshotKey",
          "detectedAt",
          "status",
          "ackedById",
          "ackedAt",
          "ackedBy",
          "resolvedById",
          "resolvedAt",
          "resolvedBy",
          "space",
          "createdAt"
        ]
      },
      "AlertList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Alert"
        }
      },
      "AlertNote": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "createdBy": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "authorRole": {
            "type": "string",
            "enum": [
              "ADMIN",
              "STAFF"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "note",
          "createdBy",
          "authorRole",
          "createdAt"
        ]
      },
      "AlertDetail": {
        "type": "object",
        "properties": {
          "alertSeq": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "id": {
            "type": "string"
          },
          "backendEventId": {
            "type": "string"
          },
          "facilityId": {
            "type": "string"
          },
          "cameraId": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "spaceId": {
            "type": "string"
          },
          "room": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "probability": {
            "type": "number"
          },
          "snapshotKey": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "detectedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "ACKED",
              "RESOLVED"
            ]
          },
          "ackedById": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "ackedAt": {
            "oneOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "ackedBy": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AlertActor"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolvedById": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolvedAt": {
            "oneOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolvedBy": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AlertActor"
              },
              {
                "type": "null"
              }
            ]
          },
          "space": {
            "$ref": "#/components/schemas/AlertSpace"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertNote"
            }
          }
        },
        "required": [
          "alertSeq",
          "id",
          "backendEventId",
          "facilityId",
          "cameraId",
          "spaceId",
          "room",
          "type",
          "probability",
          "snapshotKey",
          "detectedAt",
          "status",
          "ackedById",
          "ackedAt",
          "ackedBy",
          "resolvedById",
          "resolvedAt",
          "resolvedBy",
          "space",
          "createdAt",
          "notes"
        ]
      },
      "AlertSseData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "backendEventId": {
            "type": "string"
          },
          "alertSeq": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "facilityId": {
            "type": "string"
          },
          "spaceId": {
            "type": "string"
          },
          "cameraId": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "type": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "ACKED",
              "RESOLVED"
            ]
          },
          "probability": {
            "type": "number"
          },
          "detectedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "backendEventId",
          "alertSeq",
          "facilityId",
          "spaceId",
          "cameraId",
          "type",
          "status",
          "probability",
          "detectedAt"
        ]
      },
      "AlertUpdatedSseData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "alertSeq": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "spaceId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACKED",
              "RESOLVED"
            ]
          },
          "resolvedById": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolvedAt": {
            "oneOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "alertSeq",
          "spaceId",
          "status",
          "resolvedById",
          "resolvedAt"
        ]
      }
    }
  }
}